# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GreenLogix API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/driver/route": {
      "get": {
        "tags": [
          "driver"
        ],
        "summary": "Driver Route",
        "operationId": "driver_route_driver_route_get",
        "parameters": [
          {
            "name": "plate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plate"
            }
          },
          {
            "name": "X-Driver-Pin",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Driver-Pin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverRouteList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stops/{id}/status": {
      "post": {
        "tags": [
          "driver"
        ],
        "summary": "Stop Status",
        "operationId": "stop_status_stops__id__status_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "X-Driver-Pin",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Driver-Pin"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stops/{id}/photo": {
      "post": {
        "tags": [
          "driver"
        ],
        "summary": "Stop Photo",
        "operationId": "stop_photo_stops__id__photo_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "X-Driver-Pin",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Driver-Pin"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_stop_photo_stops__id__photo_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/driver/restriction-feedback": {
      "post": {
        "tags": [
          "driver"
        ],
        "summary": "Restriction Feedback",
        "description": "Queue a driver restriction report for admin review (T-03).\n\nFeedback NEVER auto-mutates ACTIVE_RULES; verified entries require a\nseparate admin promotion step (no auto-promotion path exists).",
        "operationId": "restriction_feedback_driver_restriction_feedback_post",
        "parameters": [
          {
            "name": "X-Driver-Pin",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Driver-Pin"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/driver/restriction-feedback/pending": {
      "get": {
        "tags": [
          "driver"
        ],
        "summary": "Restriction Feedback Pending",
        "description": "List queued driver restriction reports awaiting review (admin only).\n\nRead-only over the in-memory queue. Never touches ACTIVE_RULES.",
        "operationId": "restriction_feedback_pending_driver_restriction_feedback_pending_get",
        "parameters": [
          {
            "name": "X-Driver-Pin",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Driver-Pin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedbackItemOut"
                  },
                  "title": "Response Restriction Feedback Pending Driver Restriction Feedback Pending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/driver/restriction-feedback/{feedback_id}/verify": {
      "post": {
        "tags": [
          "driver"
        ],
        "summary": "Restriction Feedback Verify",
        "description": "Verify or reject one queued report (admin only).\n\nCalls verify_feedback only; never touches ACTIVE_RULES \u2014 admin\npromotion to routing rules is a separate manual step (no path exists).\nUnknown id returns 404.",
        "operationId": "restriction_feedback_verify_driver_restriction_feedback__feedback_id__verify_post",
        "parameters": [
          {
            "name": "feedback_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feedback Id"
            }
          },
          {
            "name": "X-Driver-Pin",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Driver-Pin"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackVerifyIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_stop_photo_stops__id__photo_post": {
        "properties": {
          "photo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Photo"
          }
        },
        "type": "object",
        "required": [
          "photo"
        ],
        "title": "Body_stop_photo_stops__id__photo_post"
      },
...
      "DriverRouteList": {
        "properties": {
          "routes": {
            "items": {
              "$ref": "#/components/schemas/DriverRouteOut"
            },
            "type": "array",
            "title": "Routes"
          }
        },
        "type": "object",
        "required": [
          "routes"
        ],
        "title": "DriverRouteList"
      },
      "DriverRouteOut": {
        "properties": {
          "plate": {
            "type": "string",
            "title": "Plate"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/StopOut"
            },
            "type": "array",
            "title": "Stops"
          }
        },
        "type": "object",
        "required": [
          "plate",
          "stops"
        ],
        "title": "DriverRouteOut"
      },
      "FeedbackIn": {
        "properties": {
          "plate": {
            "type": "string",
            "title": "Plate"
          },
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng"
          },
          "issue_type": {
            "type": "string",
            "enum": [
              "road_closed",
              "height_barrier",
              "weight_limit",
              "unexpected_ban"
            ],
            "title": "Issue Type"
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "plate",
          "lat",
          "lng",
          "issue_type"
        ],
        "title": "FeedbackIn"
      },
      "FeedbackItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "plate": {
            "type": "string",
            "title": "Plate"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "issue_type": {
            "type": "string",
            "enum": [
              "road_closed",
              "height_barrier",
              "weight_limit",
              "unexpected_ban"
            ],
            "title": "Issue Type"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending_review",
              "verified",
              "rejected"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "driver_id",
          "plate",
          "lat",
          "lng",
          "issue_type",
          "notes",
          "status"
        ],
        "title": "FeedbackItemOut",
        "description": "One queued driver restriction report (admin review read model)."
      },
      "FeedbackOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending_review",
              "verified",
              "rejected"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "FeedbackOut"
      },
      "FeedbackVerifyIn": {
        "properties": {
          "approved": {
            "type": "boolean",
            "title": "Approved"
          }
        },
        "type": "object",
        "required": [
          "approved"
        ],
        "title": "FeedbackVerifyIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StatusIn": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "arrived",
              "delivered",
              "failed"
            ],
            "title": "Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "khach_vang",
                  "sai_dia_chi",
                  "hang_hong",
                  "tu_choi"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Required and non-null when status is failed."
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusIn"
      },
      "StatusOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "arrived",
              "delivered",
              "failed"
            ],
            "title": "Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "khach_vang",
                  "sai_dia_chi",
                  "hang_hong",
                  "tu_choi"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "reason"
        ],
        "title": "StatusOut"
      },
      "StopOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "kind": {
            "type": "string",
            "enum": [
              "depot",
              "stop"
            ],
            "title": "Kind"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "window_start": {
            "type": "string",
            "title": "Window Start"
          },
          "window_end": {
            "type": "string",
            "title": "Window End"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "kg": {
            "type": "number",
            "title": "Kg"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "arrived",
              "delivered",
              "failed"
            ],
            "title": "Status"
          },
          "fail_reason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "khach_vang",
                  "sai_dia_chi",
                  "hang_hong",
                  "tu_choi"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Fail Reason"
          },
          "late_risk": {
            "type": "boolean",
            "title": "Late Risk",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "seq",
          "kind",
          "order_id",
          "lat",
          "lng",
          "address",
          "phone",
          "window_start",
          "window_end",
          "notes",
          "kg",
          "status"
        ],
        "title": "StopOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}